// src/app/shared/Model/addUser.ts
export class AddUserModel {
    user_ID: Number;
    user_First_Name: string;
    user_Last_Name: string;
    user_Email_Address: string;
    user_Primary_Phone: string;
    password: string;
    role_ID: Number;
    active: true;
    created_By: Number;
    created_Date: Date;
    modified_By: Number;
    modified_Date:Date;
    org_Id: Number
}